Refactored item and actor sheets to use popup architechure to break monolith into responsible components.
