+++ css/eventide-rp-system.css
@@ -7813,5 +7813,3 @@ body *::-webkit-scrollbar-track,
     flex: 1;
   }
 }
-
-/*# sourceMappingURL=eventide-rp-system.css.map */
