+++ module/ui/mixins/_module.mjs
@@ -1,0 +1,40 @@
+/**
+ * Sheet Mixins Module
+ *
+ * Exports all sheet mixins that provide modular functionality
+ * for the Eventide RP System sheet classes.
+ *
+ * These mixins follow the composition pattern to add specific
+ * capabilities to base sheet classes without creating large
+ * monolithic files.
+ */
+
+// Actor Sheet Action Mixins
+export { ActorSheetGearActionsMixin } from "./actor-sheet-gear-actions.mjs";
+export { ActorSheetDocumentActionsMixin } from "./actor-sheet-document-actions.mjs";
+export { ActorSheetTransformationActionsMixin } from "./actor-sheet-transformation-actions.mjs";
+export { ActorSheetDragDropMixin } from "./actor-sheet-drag-drop.mjs";
+
+// Combined Actor Sheet Mixin
+export { ActorSheetActionsMixin } from "./actor-sheet-actions.mjs";
+
+// Actor Sheet Additional Functionality Mixins
+export { ActorSheetStatusBarMixin } from "./actor-sheet-status-bar.mjs";
+export { ActorSheetThemeMixin } from "./actor-sheet-theme.mjs";
+export { ActorSheetGearTabsMixin } from "./actor-sheet-gear-tabs.mjs";
+export { ActorSheetAdditionalActionsMixin } from "./actor-sheet-additional-actions.mjs";
+export { ActorSheetFormOverrideMixin } from "./actor-sheet-form-overrides.mjs";
+export { ActorSheetContextPreparationMixin } from "./actor-sheet-context-preparation.mjs";
+
+// Combined Actor Sheet Mixin
+export { ActorSheetAllMixins } from "./actor-sheet-all-mixins.mjs";
+
+// Item Sheet Mixins
+export { ItemSheetThemeMixin } from "./item-sheet-theme.mjs";
+export { ItemSheetDragDropMixin } from "./item-sheet-drag-drop.mjs";
+export { ItemSheetCharacterEffectsMixin } from "./item-sheet-character-effects.mjs";
+export { ItemSheetActionsMixin } from "./item-sheet-actions.mjs";
+export { ItemSheetEffectGuardsMixin } from "./item-sheet-effect-guards.mjs";
+
+// Combined Item Sheet Mixin
+export { ItemSheetAllMixins } from "./item-sheet-all-mixins.mjs";
