+++ templates/actor/actor-status.hbs
@@ -4,7 +4,7 @@
     {{#if actor.flags.eventide-rp-system.activeTransformation}}
       <article class="eventide-transformation-card {{#if actor.flags.eventide-rp-system.activeTransformationCursed}}eventide-transformation-card--cursed{{/if}}"
                data-drag="true"
-               data-item-id="{{activeTransformation._id}}"
+               data-item-id="{{activeTransformationId}}"
                data-document-class="Item"
                title="{{localize 'EVENTIDE_RP_SYSTEM.Actor.Attributes.Transformation.DragHint'}}">
         <header class="eventide-transformation-card__header">
